+++ e_face_x4/app/static/service-worker.js
@@ -1,4 +1,4 @@
-self.EFACE_SERVICE_WORKER_VERSION = '2.21.57'
+self.EFACE_SERVICE_WORKER_VERSION = '2.21.61'
 self.addEventListener('install', () => self.skipWaiting())
 self.addEventListener('activate', event => event.waitUntil(clients.claim()))
 self.addEventListener('push', event => {